Remove broken web-mode support
`nerd-icons--web-mode-icon` is supposed to lookup the correct icon for a
web-mode buffer, but it doesn't work in practice:

1. It's requires the web-mode buffer to be current (it relies on
   buffer-local variables) but that's rarely the case (e.g., when called
   from `nerd-icons-icon-for-mode`).

2. `nerd-icons-icon-for-mode` caches the icon, so the chosen icon will
  be correct the first time only (and then, only correct subject to 1).

Fortunately, there's a better option: use `nerd-icons-icon-for-buffer`.
That will try to derive an icon from the file-name first before falling
back on the mode.

This patch removes this support as it's broken and there's a working
alternative.