Make `nerd-icons-auto-mode-match?` work with tree-sitter modes
`nerd-icons-auto-mode-match?` checks if a buffer's mode matches it's
automatically set mode. Unfortunately, the function in `auto-mode-alist`
may not match the actual mode set on the buffer because:

1. Mode-remapping can select a different mode.

2. The function can actually be an arbitrary function that selects the
mode based on some other criteria. This is most often used in
tree-sitter modes to automatically select between the tree-sitter
version of the mode and the non-tree-sitter version.

The fix here is to check the `set-auto-mode--last` variable (when
available) against the buffer's current mode. Unfortunately, this is
abusing a private variable.

This patch also removes the FILE argument from
`nerd-icons-auto-mode-match?` because this function applies to the
current buffer so there's no real reason to match on anything but the
current file.